Changes colours of site for more contrast / readability
### What does this PR do?

* Changes the website to use more contrasting colour for readibility with
the purple.

* If you use the current colour codes against Infinima on the docusaurus
site, it is failing the "contrast check".

* Adds `yarn run website:dev` testing the site / looking for active
  changes instead of building

### Screenshot/screencast of this PR

See below for updated colour codes that pass the docusaurus Infinima
contrast check.

### What issues does this PR fix or reference?

The readibility of the purple against the gray background / black is too
light (and sometimes too dark) for me to be able to read it clearly.

N/A (let me know if I should open up issues for anything minor like
this!)

### How to test this PR?

<!-- Please explain steps to reproduce -->

Run `yarn run website:dev`

Signed-off-by: Charlie Drage <charlie@charliedrage.com>