chore: bump dockerode/ssh2 to their latest version Before we had some issues so we were using an old version of ssh2 but now, this is working
disable also electron builder rebuilding of native modules
as it's failing on optional dependency of ssh2 which is
cpu-features.

So while it's an optional dependency, developers are seeing a huge
compilation error everytime. By disabling the rebuilding, we
don't try to compile the optional dependency, then no error is
displayed

Change-Id: I6ae3537e8b04211f64461c2e039ec1b81f92ff0d
Signed-off-by: Florent Benoit <fbenoit@redhat.com>