fix: 'empty screen' icons
Fix each of the four empty screens to use the correct SVG icon and reduce code duplication via EmptyScreen.

Switched the order of the if statement in each of the lists because I thought this is more natural, and avoids passing array into each screen to confirm if it is empty.

The volumes empty screen does not appear before or after this change because filtered.length == 1 when there are no visible volumes, will continue investigating via separate PR.

Signed-off-by: Tim deBoer <git@tdeboer.ca>