fix: Remove padding from SVG icons
Our SVG icon components all contained padding, which made it harder to reuse them elsewhere at the correct size. This change removes the padding and reduces the size on the empty screen to match (and correctly match FA icons).

I confirmed with design that the bigger/unpadded icons are the correct size to use for the nav, so I removed padding from the remaining icon SVGs (dashboard, extensions, settings) as well, and a slight increase to the text padding (mx) was requested.

This also has the benefit that if we release PR #1058 the extension icons would be at most the same size and not bigger than ours.

Signed-off-by: Tim deBoer <git@tdeboer.ca>