chore: Differentiate icons vs status icons
Last week I (inadvertently) created reusable icon SVGs (pods, containers, etc) with identical names as some existing status icons. To make it clear which is which I'm renaming them as follows:

*Icon: a reusable representation icon for something, typically just an SVG. Simple properties, e.g. whether the icon should be solid (filled). Usable much like Fort Awesome.
*StatusIcon: a reusable status/state icon, typically a div that may include color, a border, etc. Properties like status, in-use state.

ImageStatusIcon already existed, this renames the equivalent Container and Volume components to match. Pod(Status)Icon is only an SVG and already mostly redundant, but since it requires a bit more work to replace I'll leave it to a future PR.

Signed-off-by: Tim deBoer <git@tdeboer.ca>