nit: change window creation background colour to dark
### What does this PR do?

Changes the background colour from the default white, to a darker shade
(uses the same dark background colour within Podman Desktop).

This prevents the white "flicker" that sometimes happens when loading,
or when clicking on the application icon in the tray milliseconds before it fully loads.

### Screenshot/screencast of this PR

<!-- Please include a screenshot or a screencast explaining what is doing this PR -->

N/A

### What issues does this PR fix or reference?

<!-- Please include any related issue from Podman Desktop repository (or from another issue tracker).
-->

N/A

### How to test this PR?

`yarn watch` and it'll have no more white "flicker" before loading (may
have to try multiple times to try and see the background).

<!-- Please explain steps to reproduce -->

Signed-off-by: Charlie Drage <charlie@charliedrage.com>