chore: Start our own palette
Mairin is defining a Tailwind palette to use across all designs. For example, instead of a checkbox being '#64ad6c' it'll just be 'green-400' which we can use directly in the code.

This first step adds several colours we haven't used yet (charcoal, dusty purple, fuschia, and sky), and overrides the color values for two that we were minimally using (green and red). I've also cleaned up the references to 'error red' to all be red-500 and one instance of 'started' green to be green-400. There are a couple more references to a green for started state in the status icons, but I have no cleaned these up since there is already a PR open.

Mairin has already defined two other colours (gray and purple), but these are used more extensively (and hence require more review) so I've left them to a future PR. We also need to review our use of other colours (zinc/violet/indigo) and either replace them or add these to the palette. After this, we should switch these colours from extending to replacing Tailwind's set, which will make sure all future work sticks to the approved palette.

Signed-off-by: Tim deBoer <git@tdeboer.ca>