chore: Simplify image inUse
I noticed (while working on a related UI PR) that inUse is always false within the image details page. This is because the page is not passing through any container information.

Instead of making the details page subscribe to containers and pass the info through to image-utils.ts, it seems a simpler change to use the Containers count similar to how volume-utils.ts uses UsageData?.RefCount. Is this change safe (i.e. this counter was just missed during implementation)? The identical images are marked inUse for me.

Signed-off-by: Tim deBoer <git@tdeboer.ca>