fix: delete enablement for engines that support Container count
Partial fix for issue #1504. See previous attempt in PR #1437.

The problem with delete action enablement and incorrect icon will still exist in the ImageDetails page for container engines that don't support a container count, but this will fix the problem for those that do, and avoid the extra processing in that path. I don't know how others feel about partial fixes, but IMHO this solves the problem immediately and safely for one class of engines.

Couldn't help myself, I also fixed the odd variable names.

Signed-off-by: Tim deBoer <git@tdeboer.ca>