fix: Tabs not selected in Container and Image Details pages
I noticed that some of the tabs on the container and image details pages weren't correctly showing as selected when they should be (e.g. see the pod details summary screenshot in #1554). I tracked this down to some missing encodeURI() calls that the other pages were doing, which caused the pf-m-current style to be missing.

I started to fix this, but it seemed like a lot of copy/paste code so I pulled a baseURL out into a variable... then I thought it's even better to use the router and avoid all calls to encodeURI()... then went full Svelte component. In the end this is a lot more cleanup than necessary to fix a simple bug, but the result is far less code, ensured consistency, and easier to update/maintain.

Signed-off-by: Tim deBoer <git@tdeboer.ca>