feat: Introduce ErrorMessage component
Creates a new ErrorMessage component with two properties: an error message, and a boolean to control whether it should be shown as an icon with tooltip or full text. In full text mode the same icon is shown (for consistency and to reinforce our error icon).

Initial PR for issue #1365. I wanted to get feedback/approval on the component and initial use before I try to update more errors.

Each mode is used in two places so far: the icon is used in the container list and details (for containers that fail to start), and the full text mode when running an image and installing extensions. For the last one I tweaked the layout and padding a little to make it more consistent with other UI.

Signed-off-by: Tim deBoer <git@tdeboer.ca>