chore: Use error message component
This commit updates all remaining error messages to use the error message component.

I've added {$$props.class} to the component so that styling can be applied directly to the component. I've left if statements around the use in cases where we were not clearing the space unless an error existed.

The only remaining use of 'red-' is DeployPodToKube:342, but this is a status component not an error message.

Signed-off-by: Tim deBoer <git@tdeboer.ca>