fix: Remove nav page subtitles
The subtitles for the list views said "Hover over X to view action buttons; ...", which is out of date with the current UI. We started to have a discussion about what to change it to on Thursday afternoon, but then decided that our UI is obvious enough and we prefer the clean look.

I reviewed the other places that had subtitles, and all of these either used &nbsp to keep the layout consistent, or had subtitles that didn't add anything significant beyond the title, so I am removing the ability to add subtitles entirely (git history remains if we ever need it again).

Tweaked nav page layout so that the lower padding remains the same size with or without the search bar.

Signed-off-by: Tim deBoer <git@tdeboer.ca>