Improved responsiveness: container eng name column
This improves the responsiveness of the left most name column of
container engine cards that appear in Settings > Resources, particularly
for names that are very long. The column has a min width of 150px and
maxes out at 200px at which point it creates line breaks at word
boundaries. This commit also helps the images for each engine/orchestrator
maintain their aspect ratio in wrapping scenarios.

Signed-off-by: Máirín Duffy <duffy@redhat.com>