Use short id/name for pods and volumes, fix layout
The pod id and volume name are long strings, and often cause the action buttons to be pushed off the right of the screen. Using shortId and shortName fix the problem.

Both the left details and right action panel were set to w-full, which would often mean that the details would get wrapped even though there was enough space. Removed the w-full from the action panels so that they never take more space than the width of the buttons.

Signed-off-by: Tim deBoer <git@tdeboer.ca>