nit: ignore kind in .gitignore when running tests
### What does this PR do?

When running tests, sometimes the kind tests creates a binary in the
podman-desktop folder.

`kind` should be added to `.gitignore` so when we run `yarn test`, we
don't have to remove `kind` eveyr time

### Screenshot/screencast of this PR

<!-- Please include a screenshot or a screencast explaining what is doing this PR -->

N/A

### What issues does this PR fix or reference?

<!-- Please include any related issue from Podman Desktop repository (or from another issue tracker).
-->

### How to test this PR?

<!-- Please explain steps to reproduce -->

Run `yarn test`, even if there is a `kind` in the source dir, gitignore
will ignore it.

Signed-off-by: Charlie Drage <charlie@charliedrage.com>