feat: Update design of Prune dialog
Expose the message box API to the renderer (but not extensions) and switch the Prune dialog to use it. This changes the prune dialog to match current design guidelines.

Although we may still need to update the Modal component to the new design, this simplifies the prune code as well.

Partial fix for issue #1987.

Signed-off-by: Tim deBoer <git@tdeboer.ca>