chore: Clean up src/lib folder
All detail pages (pods, images, volumes) are in their own respective folders except for ContainerDetails. All preferences are in the /preferences folder except for the ExtensionList. This just moves each of these pages into the normal locations that we've been putting everything else, and fixes the references to them.

No functional/UI changes, this just cleans up the renderer/src/lib folder, which makes it easier and more consistent to find things.

Signed-off-by: Tim deBoer <git@tdeboer.ca>