chore: extension events
The main extensions loader was firing an extension-removed event, but all other extension lifecycle events (including a second extension-removed) were fired within the renderer level. This adds an extension-started and -stopped event to the loader, and migrates all the listeners to use these events instead. IMHO it is simpler and safer down the road when a change event comes from the thing doing the change, and isn't reliant on multiple UIs triggering the change to also fire an event.

The comment in containers.ts said it was listening to extension started and stopped events, but the code was only listening to started. I tested the code and it is getting several provider-change events when extensions change, so I've fixed the comment instead of the code.

extensions.ts was listening for the UI extension-removed event even though there was one from the loader, so I've made it consistent.

This is a fix towards issue #1810; having consistent events will make it easier to fix the UI actions.

Signed-off-by: Tim deBoer <git@tdeboer.ca>