feat: Fixed, simple navigation bar
I had no clue I'd end up working on this today, but after a long day I thought I'd try one topic that came up earlier and it turned out to be pretty straight-forward.

This changes the main navigation to be a fixed width icon bar that is always visible.

- Removed all titles and counter badges. Added tooltips.
- There were several redundant divs and unused code removed. I tried scanning all other classes to fix inconsistencies.
- Extensions were expandable before - I made the call to remove the Extensions icon and always include them.
- I made the Settings title have slightly less top padding so that it is at the same position as other titles like Images/Container list.
- Removed Exit Settings link.

Commit redone to:
- Fix branch conflict due to a different PR.
- Add separator bar between regular icons and extensions.
- Support clicking on Settings a second time and returning to the previous page.

Fixes issue #2167 and issue #1648.

Signed-off-by: Tim deBoer <git@tdeboer.ca>