fix: empty screen copy/paste background color
Fixing the background color of the copy/paste field.

While I'm at it, I'm also switching the EmptyScreen to correctly use Svelte's ability to pass down style/class (as we do in most other components), fixing the one use of it in the resource settings, along with typos in the message.

Signed-off-by: Tim deBoer <git@tdeboer.ca>