feat: Sort extensions
Extensions should be sorted in both the Settings > Extensions page and in the child list by their display name. This incidentally fixes issue #2130, which was caused by the children being out of order alphabetically.

While fixing this I reduced the font weight of the extension child items to match the Setting > Preferences children in the navigation (and verified with mairin).

Signed-off-by: Tim deBoer <git@tdeboer.ca>