feat: Remove extension from its own page
Today you can remove extensions from the main page, but not the child page specific to that extension. This adds an action to remove user extensions (bumping the user back to the main page), and a comment for built-in extensions that you can't remove them.

We haven't done a proper redesign for these pages, but it seems likely we'll need this action on every page and it resolves the first part of issue #1392 (hard to find delete action).

Signed-off-by: Tim deBoer <git@tdeboer.ca>