fix: Push image modal styling
The push image modal had some slightly unique styling before, but changing the palette has tipped it well over the edge into unusable territory. This fix updates all modals to the same transparency as the newer dialogs, and reuses the same outer divs as the dialogs for pushing images.

Long term I would prefer to embed the outer box and title bar in the Modal component for consistency, but since there are 8 dialogs all with their own styling today that is a bigger change that I don't want to attempt now.

Signed-off-by: Tim deBoer <git@tdeboer.ca>