feat: UI support for markdown preference descriptions
Allow preference descriptions to use markdown, and update the telemetry description to include a proper link to our privacy statement. This would be _so_ much better with a fix for issue #2230, but I don't think that's a blocker since this is shorter, less ugly, and already easier to copy the link (Copy Link popup vs previous requirement to select the right text and copy).

Signed-off-by: Tim deBoer <git@tdeboer.ca>