chore: Add image pull validation tests
Add tests to validate that the button to pull an image/extension is disabled with whitespace and enabled when there is a valid entry.

In all three cases it required exposing the image name from the component. For PullImage I had to simplify the imageToPull (remove unused registryServerUrl) and trigger initial validation. Otherwise, the same pattern is used in all three places and tests.

Signed-off-by: Tim deBoer <git@tdeboer.ca>