chore: Add button enablement test
Adds tests to confirm that the extension start/stop/remove buttons are enabled correctly based on the state of a mock extension.

The page itself did not need any changes, but I had to add ? to extensions.ts to avoid nulls. Although it's not directly related I was also frequently hitting a timing issue with TelemetryService, so I added a check there too.

Signed-off-by: Tim deBoer <git@tdeboer.ca>