feat: Remove extension from its own page (#2242)
* feat: Remove extension from its own page

Today you can remove extensions from the main page, but not the child page specific to that extension. This adds an action to remove user extensions (bumping the user back to the main page), and a comment for built-in extensions that you can't remove them.

We haven't done a proper redesign for these pages, but it seems likely we'll need this action on every page and it resolves the first part of issue #1392 (hard to find delete action).

Signed-off-by: Tim deBoer <git@tdeboer.ca>

* chore: Add button enablement test

Adds tests to confirm that the extension start/stop/remove buttons are enabled correctly based on the state of a mock extension.

The page itself did not need any changes, but I had to add ? to extensions.ts to avoid nulls. Although it's not directly related I was also frequently hitting a timing issue with TelemetryService, so I added a check there too.

Signed-off-by: Tim deBoer <git@tdeboer.ca>

---------

Signed-off-by: Tim deBoer <git@tdeboer.ca>