fix: Remove extra margin from resource creation
There is an extra margin between the labels and fields on the resource creation pages, and this removes it. I also made the padding between properties just slightly smaller. This is enough to make the entire page just a little shorter and fix issue #2271.

This made it obvious that sliders (e.g. for Podman CPU creation) were missing a top margin, so I added that.

Fixes #2271.

Signed-off-by: Tim deBoer <git@tdeboer.ca>