fix: Incorrect pod status
ContainerGroupIcon was just a PodIcon, setting the status to running or stopped based on whether all containers in the group were running or not. Pods have a degraded state where only some of the containers are running.

Instead of trying to make this icon more complicated, this change falls back to directly using the PodIcon and the status of the container group. There are three kinds of groups, with the status now set in container-utils.ts:
- standalone containers - set to the container status
- pods - set to pod status
- compose groups - wait until the group is created, then use the same logic as before

Signed-off-by: Tim deBoer <git@tdeboer.ca>