fix: Avoid creating duplicate registries
The fix to disable the Login button is simple: use a loggingIn flag that disables the Login buttons while it is processing.

However, while testing this I kept getting distracted by the fact that the user/password/cancel controls were not vertically aligned and the show/hide password wasn't centered in the password box. I removed most of the top/bottom vertical padding and used items-center to center all of the rows vertically.

Signed-off-by: Tim deBoer <git@tdeboer.ca>