fix: rebase z-ordering changes
This commit rebases changes proposed by @mairin onto the current main. Goodness is hers; if anything is wrong it's my fault.

This also changes the tooltip 'right' option (verified that it's not used elsewhere) so that the tooltip is more right than bottom-right, which matches tooltip placement in Slack and avoids ordering issues with the status bar.

Signed-off-by: Tim deBoer <git@tdeboer.ca>