chore: Clean up remaining PF forms
After PR #2604 I noticed that we were only using PatternFly form classes in a couple remaining places. Maybe it's just me, but I thought this was a good time to remove this use, remove overrides, and fix minor issues with formatting. This should give us the same look with one less dependency on PF.

Signed-off-by: Tim deBoer <git@tdeboer.ca>