fix: Only show status bar when kubeconfig exists
Moved the initialization of the status bar to when the kubeconfig updates, and added a dispose when it is deleted. As a result, the status bar will not appear for anyone who is not using Kubernetes on their machine, will appear the moment they do, and will disappear if kubeconfig is ever deleted. This behaviour now matches the corresponding system tray menu.

Signed-off-by: Tim deBoer <git@tdeboer.ca>