refactor: remove adding new registry modal
### What does this PR do?

When implementing a registry feature, I noticed we have a lefover modal
that hasn't been updated in a while.

It also does not include any reccomended defaults (docker hub, quay,
etc.)

I believe this should be removed in favour of adding a registry through
the registries section rather than it's own separate (and separately
maintainable) modal.

I've removed it in favour of having the single button "manage
registries" and clicking it will take you to the page to add a registry.

### Screenshot/screencast of this PR

<!-- Please include a screenshot or a screencast explaining what is doing this PR -->

See screenshot of modal that doesn't seem to be used:

### What issues does this PR fix or reference?

<!-- Please include any related issue from Podman Desktop repository (or from another issue tracker).
-->

References: https://github.com/containers/podman-desktop/issues/1653
I fould the modal while implementing.

### How to test this PR?

<!-- Please explain steps to reproduce -->

Button should not be there anymore.

Signed-off-by: Charlie Drage <charlie@charliedrage.com>