chore: Move preferences search into header
Follow up from PR #2863: the search bar for preferences should be in the header and always visible.

I started to add a search capability to SettingsPage similar to NavPage, but I don't think the extra code and complexity is worth it when this is the only page that has search within Settings. This just adds a slot to the header, uses it in preferences, and makes the necessary formatting changes.

Minor tweak to merge pt/mt + pb into a py.

Signed-off-by: Tim deBoer <git@tdeboer.ca>