fix: overflow-auto in NavPage
Moves overflow-auto 'up' to the NavPage so that all child content automatically
scrolls as necessary. Moved top of search bar padding to bottom of title - this
helps child pages without search so that there is a little padding under the
title (when scrolling the content doesn't look like it hits the title).

Most child content divs require 'min-w-full and h-fit' so that the form stays
full width but only as large vertically as necessary, made this consistent.

Signed-off-by: Tim deBoer <git@tdeboer.ca>