fix: only show empty screen when no pods
The 'no container engine' empty screen is a little off here since Kube pods can be
detected, or you could install a Kubernetes provider instead of a container engine.
Doubly so since pods can be picked up via the Kube context even if there are no
Kubernetes providers installed.

I tried to come up with something more clever, but checking for other providers or
changing the text of the screen seems much more complicated and I still couldn't
find a path that would cover every case or be a significant improvement. The
simplest solution seems to be just not showing the screen if we've found any pods.

Signed-off-by: Tim deBoer <git@tdeboer.ca>