feat: background colors and FormPage
There are 3 pages that have lighter content backgrounds: RunImage, DeployPodToKube,
and Help. After fixing the header and scrollbars globally, these screens look odd
because the lighter background doesn't cover the entire content area when the screen
is large.

I thought these screens were incorrect compared to the others, but after checking
with @mairin these three are actually more recent designs. The intention is for all
'task pages' (Pull Image, Build Image, etc) to have the same lighter background, and
be similar to the Details pages with breadcrumbs, close button, and potentially icons
or other things in the future.

The only fix for the background color is to customize NavPage's content background -
or just define these as a new FormPage component. Since I was going to create this
component anyway I added some basic features: the same breadcrumb and close button as
we have in the details pages.

So to sum up this:
- fix: Fixes the problem with incomplete background color.
- chore: Defines common FormPage for these pages.
- feat: Adds breadcrumb and close button to these pages, matching what we have in
  the Details pages.

Future work (follow-up PRs):
- Switch the other task pages over.
- Wait for full design and add icons or other features.

Fixes #2976.

Signed-off-by: Tim deBoer <git@tdeboer.ca>