fix: fail extension load when manifest is incomplete
This changes analyzeExtension() to reject the Promise if there is an error
loading an extension (no manifest or missing required attributes), instead
of returning undefined. The main reason for the change is to allow error
messages to be propagated back to the client (e.g. UI).

The remaining changes are handling this in all the places that call
analyzeExtension(). IMHO most are more readable and handle error cases
cleaner after this change, but for lines 267/272/286 I tried several
options and couldn't find anything better than adding
'.catch(() => undefined))).'

Fixes #2871.

Signed-off-by: Tim deBoer <git@tdeboer.ca>