fix: website check targets
It was only possible to check the much bigger top-level targets...

But it should be possible to format and lint also the website dir.

Note: eslint resolve project includes from the current directory.
So it'll look for website/website, workaround by running from root.

Ideally it would be possible to use the current project directory,
but that is only avaiable from eslintrc.js and not eslintrc.json

Signed-off-by: Anders F Björklund <anders.f.bjorklund@gmail.com>