chore: refactor stores
A lot of stores were having the same pattern with a lot of copy/paste.

Refactor by introducing an EventStore class, allowing to update, fetch,
get details on events that triggered the update

It will help to build troubleshooting of these stores as they'll stick
to a same design.

Also all stores following this type are now registered in a global store
to fetch/investigate/analyze all of them.

From user point-of-view, it should work now as before.
Nothing has changed in the stores, they should be updated in the same way

part of https://github.com/containers/podman-desktop/issues/3053
Signed-off-by: Florent Benoit <fbenoit@redhat.com>