nit: change wording from started to enabled / etc.
### What does this PR do?

For providers it makes sense to have starting / stopping / stopped.

For extensions, it should be enable / disable / enabling / disabling
since they are extensions that you turn off and on, not something that
is "running" in the background.

Using the starting / stopping / stopped terms for extensions is
confusing.

### Screenshot/screencast of this PR

<!-- Please include a screenshot or a screencast explaining what is doing this PR -->

https://github.com/containers/podman-desktop/assets/6422176/e8914923-e19b-4dd2-bad7-ab35dd7251fe

### What issues does this PR fix or reference?

<!-- Please include any related issue from Podman Desktop repository (or from another issue tracker).
-->

Closes https://github.com/containers/podman-desktop/issues/2534
Closes https://github.com/containers/podman-desktop/issues/2379

### How to test this PR?

<!-- Please explain steps to reproduce -->

Go to extensions page, you should now see enable / disable rather that
start / stop.

Signed-off-by: Charlie Drage <charlie@charliedrage.com>