feat: switch more pages to formpage
Changes almost all of the remaining pages to FormPages, using the existing icons
that are used in the actions to open the form and the form button:
- Build image
- Pull image
- Play Kube YAML
- Create Pod from containers

Adds a new 'actions' slot for the PullImage page's Manage registries button.
Adds 'missing' pull image action to the button to match other pages.
Makes Kube play icon resizable so that a larger form can be used in the header.

Svelte has no API to test slots directly (see below), so a FormPageSpec component is
added solely to be able to test that the icon/actions/content slots work correctly.
The actions section is also tested by confirming the pull image registry button.

https://github.com/testing-library/svelte-testing-library/issues/48
https://stackoverflow.com/questions/60771586/testing-svelte-components-that-use-slots

Signed-off-by: Tim deBoer <git@tdeboer.ca>