fix: move actions to title row
Mairin clarified the design - the header should essentially be two rows:
- breadcrumb row with breadcrumb, page name, and close action
- titlebar with actions on the right

i.e. the top row is only for navigation, the second row gives the title
and actions for the page. The only real impact is that the 'Manage repos'
button on the Pull Image page will move from the top line to under the X.

The same change is intended to be made to the DetailsPage, but we need to
decide how to deal with the container CPU/memory use first.

Signed-off-by: Tim deBoer <git@tdeboer.ca>