fix: limit tabbing within messagebox
The MessageBox wasn't doing anything special with tabbing, so if you
tabbed it would cycle through everything in the background as well -
e.g. every action on every container if you were on the Containers page.

This uses the same code as Modal to limit tabbing to what's visible
within the MessageBox.

Fixes at least part of #2356.

Signed-off-by: Tim deBoer <git@tdeboer.ca>