feat: Close details page if object is deleted
Added a close() function on DetailsPage so that child pages do not need to import
lastPage and router. This means they do need to bind DetailsPage, but that seemed
like simpler logic/less dependencies to me.

Each page just checks if they had the object before and can no longer find it.

Once we've done this, the router.goto() when the user triggers deletion via each
action is unnecessary and can be removed. If the user triggers the action from a
page like Images the goto wasn't doing anything, and if they are in Image Details
this new mechanism will do the same thing and revert to the previous page.

The pods page was unsubscribing the listener, but I've switched it to returning
the unsubscribe so it's handled by Svelte - no need for onDestroy() or imports.
The other three pages weren't unsubscribing the listener, so I've fixed that by
adding a return statement.

This change was meant to handle external deletion, but once it is in there is no
longer a need for any of the delete actions to use goto() and the
redirectAfterDelete from PR #2963 can be cleaned up. We also get a minor bug fix
b/c if you went from Containers > Pod Details and delete it we'll now go back to
Containers instead of Pods.

The PodActions test will fail because it is checking for the explicit redirect
when a pod (that's not registered) gets deleted. Putting this up to start review
of the rest and decide if we just delete the test or these is something else I
can do.

Fixes #3217.

Signed-off-by: Tim deBoer <git@tdeboer.ca>