fix: settings navigation bar resizing
The Settings navigation bar is unnecessarily resizing slightly as
you make the window wider or narrower. Worse, Authentication is oddly
set to disappear at widths smaller than 768px.

Set a fixed size in the middle of the current sizes and made
Authentication styling match other items.

Related to/found when fixing issue #3230, but not directly related.

Signed-off-by: Tim deBoer <git@tdeboer.ca>