fix: update tab hover
Update bottom border hover to charcoal-100, using class:hover syntax so that
the hover doesn't override the active border color (confirmed with Mairin).
I also moved the px-4 to the href and added a py-2. This doesn't affect the
visuals at all, but gives a bigger area that you can click on to activate
the tab.

Signed-off-by: Tim deBoer <git@tdeboer.ca>