feat: kubernetes icon
I was never happy with the 'server' icon for Kubernetes context and wanted to
put the new icon contributions to good use. I spent a while finding a path
to convert the official Kubernetes SVG to woff2 - and my icon looked awful.

I tried the entire path again today with the same ugly result, until I
realized it's because all custom icons are set to italics :), which I think
is wrong so I have also fixed it here.

Signed-off-by: Tim deBoer <git@tdeboer.ca>