chore: pd buttons
We currently use PatternFly classes like pf-c-button, pf-m-primary, pf-c-button__icon,
and pf-m-start to style our buttons. Although they're powerful and we have a similar
design, they tie us to PatternFly and its choices, and weren't really meant for
desktop apps or dark mode. There are a few minor differences from our original design:
- slightly more button rounding
- in dark mode the hover should be lighter, not darker
- disabled buttons should be lighter

This change uses @layer to define our own CCS classes, then changes the lists, most
form pages, and the messagebox to use it. This is just a start to use the buttons in
some core areas; after this PR it is very easy work to search & replace all other
instances in a batch or two.

Why our own CCS classes vs a component? For really lightweight styling this is
preferable, and in this case makes it an easy search & replace to switch. I've
used the pd- prefix to keep our own custom classes easily recognizable and
searchable. See more here:
https://tailwindcss.com/docs/reusing-styles#extracting-classes-with-apply

Signed-off-by: Tim deBoer <git@tdeboer.ca>