chore: button type & icon, spinner property
Responding to feedback:
- Changed default button type to primary and give third option a name ('link')
- Made spinner support a boolean property to keep it simpler
- Changed button icon to a component using the same logic as EmptyScreen

Updated all uses of Button and tests to support the changes.

Note the one ugliness on Button.svelte:18 - we need to know when an icon like
PodIcon is used because it has some required properties, and setting them on
other components would produce a log warning like `<KubePlayIcon> was created
with unknown prop 'solid'`. This isn't great but I couldn't find a simpler way
to do it and IMHO still much better overall.

Signed-off-by: Tim deBoer <git@tdeboer.ca>