chore: remove invalid test
This test checked whether the action triggers the router when it is configured
to do so. But now the router is triggered automatically and the action is not
involved. I'm not sure what/if there is a right unit test, but since this is
the wrong one I am removing that part of the test.

Signed-off-by: Tim deBoer <git@tdeboer.ca>