chore: pf-empty screen
We have three places where we were using the PF pf-c-empty classes: the EmptyScreen
and two places where for I'll assume for historical reasons we weren't using the
EmptyScreen. (and in the case of ContainerDetailsTerminal was also using a different
terminal icon than all other places)

This replaces the EmptyScreen PF classes with our own, simpler styling that gives
the identical look, and updates both of the other instances to use the EmptyScreen
component. The connection creation screen needs a button under the text, so I
added a slot to the empty screen.

Signed-off-by: Tim deBoer <git@tdeboer.ca>