fix: default last page
Breadcrumbs are getting cleared after sleep/hibernate, or when making some changes
during yarn test. If one of the Details or FormPages was left open, this means
the 'last' page is lost/unknown, which causes an HMR error.

I've tried to research how Svelte stores are persisted and Electron stores cache,
and there isn't much useful. It looks like at some point we may want to persist
the breadcrumbs (and potentially other stores) like this:
https://stackoverflow.com/questions/56488202/how-to-persist-svelte-store

For now, that seems overkill to me - it's not critical that the back-link is correct
as long as it doesn't HMR, and we can always persist later. This commit just sets a
default last page to avoid the error.

Fixes #3334.

Signed-off-by: Tim deBoer <git@tdeboer.ca>